branched to version1
corrected the webpack files so that grunt serve now works. Adjusted
the required version of webpacks needed to a lower level. (Used google
to figure out how to debugg errors going on when npm install was ran)

Required the events.js file in the index.js folder so that signup
services work correctly.